Create Calculator.html
The web page should contain following input elements and apply the specified Constraints:

 

Label Name

Element Name

Constraints

Input1

input1

This element is to get the first input. Type should be 'number'

Input2

input2

This element is to get the second input. Type should be 'number'

Select Operation

operation

A drop down list contains the following values: Select.. , ADD, SUBTRACT, MULTIPLY and DIVIDE.

Set these values as its option tag text and option tag's 'value' attributes

 

submit

An input field with image as type.

 

reset

An input field with image as type.

 

Consider the images are in the current folder and

Use “calculator.jpg” as a header image.
Use “calc.jpg” as calculation button and “reset.jpg” as reset button.

Apply following styles to the attributes: Do not use CSS . 

The heading should be done using the font color as 'blue' and with font size as '20'. (Use <font> tag)
The height and width of the images with the name “submit” and “reset” should be '80'.
The height and width of the calculator image should be '300' and '400'.